+++ code/icarus/taskmanager.h
@@ -70,7 +70,7 @@ public:
 	// Overloaded new operator.
 	inline void *operator new( size_t size )
 	{	// Allocate the memory.
-		return IGameInterface::GetGame()->Malloc( size );
+		return IGameInterface::GetGame()->Malloc( static_cast<int>(size) );
 	}
 
 	// Overloaded delete operator.
@@ -114,7 +114,7 @@ public:
 	// Overloaded new operator.
 	static void *operator new( size_t size )
 	{	// Allocate the memory.
-		return IGameInterface::GetGame()->Malloc( size );
+		return IGameInterface::GetGame()->Malloc( static_cast<int>(size) );
 	}
 	// Overloaded delete operator.
 	static void operator delete( void *pRawData )
@@ -180,7 +180,7 @@ public:
 	// Overloaded new operator.
 	inline void* operator new( size_t size )
 	{	// Allocate the memory.
-		return IGameInterface::GetGame()->Malloc( size );
+		return IGameInterface::GetGame()->Malloc( static_cast<int>(size) );
 	}
 
 	// Overloaded delete operator.
